import-advisories/pyproject.toml
[tool.poetry]
name = "badvisories"
version = "0.1.0"
description = ""
authors = ["Arpad Ray <arpadr@google.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Scrapy = "^2.6.1"
typer = "^0.4.1"
scrapyscript = "^1.1.5"
python-frontmatter = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
pylint = "^2.14.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
